nat logic factored out
